Additional documentation and cleanup
