+++ app/src/main/kotlin/Navigation.kt
@@ -1,12 +1,17 @@
 package com.adsbynimbus.android.sample
 
+import android.app.AlertDialog
 import android.content.Context
 import android.os.Bundle
+import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.view.inputmethod.InputMethodManager
 import android.widget.LinearLayout
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.widget.TextViewCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
 import androidx.navigation.NavGraph
@@ -14,11 +19,16 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavType
 import androidx.navigation.createGraph
 import androidx.navigation.findNavController
+import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.fragment
+import androidx.navigation.ui.setupWithNavController
+import androidx.preference.PreferenceManager
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.adsbynimbus.android.sample.databinding.ActivityNavigationBinding
+import com.adsbynimbus.android.sample.databinding.CustomDialogBinding
 import com.adsbynimbus.android.sample.demand.*
 import com.adsbynimbus.android.sample.mediation.GAMDemoFragment
 import com.adsbynimbus.android.sample.rendering.AdManagerFragment
@@ -112,6 +122,65 @@ interface Destination : NavItem
 @JvmInline value class Unity(val demo: String): Destination { override val text: String get() = demo }
 @JvmInline value class Vungle(val demo: String): Destination { override val text: String get() = demo }
 @JvmInline value class Header(val name: String): Decoration { override val text: String get() = name }
+
+class NavigationActivity : AppCompatActivity() {
+
+    inline val inputMethodManager get() = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+    inline val title get() = getString(R.string.main_title)
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        ActivityNavigationBinding.inflate(layoutInflater).also { setContentView(it.root) }.apply {
+            with(supportFragmentManager.findFragmentById(R.id.nav_host) as NavHostFragment) {
+                navController.graph = navController.appGraph()
+                toolbar.setupWithNavController(navController)
+
+                navController.addOnDestinationChangedListener { controller, dest, args ->
+                    val isMainNavDestination = dest.route in arrayOf("Main", "Internal")
+                    val isAdDemo = args?.containsKey("item") ?: false
+                    headerTitle.apply {
+                        text = when {
+                            isMainNavDestination -> title
+                            isAdDemo -> args?.getString("item")
+                            else -> dest.route
+                        }
+                        gravity = if (isMainNavDestination) Gravity.BOTTOM or Gravity.CENTER_HORIZONTAL else Gravity.BOTTOM
+                        TextViewCompat.setTextAppearance(
+                            this,
+                            if (isMainNavDestination) R.style.MainTitle else R.style.Title
+                        )
+                    }
+                    headerSubtitle.apply {
+                        text = dest.label ?: controller.previousBackStackEntry?.destination?.route
+                        gravity = if (isMainNavDestination) Gravity.TOP or Gravity.CENTER_HORIZONTAL else Gravity.TOP
+                    }
+
+                    inputMethodManager.hideSoftInputFromWindow(root.windowToken, 0)
+                    toolbar.title = null /* Clear title to prevent issues with UI */
+                }
+            }
+        }
+        PreferenceManager.getDefaultSharedPreferences(this).apply {
+            registerOnSharedPreferenceChangeListener(SettingsListener)
+            initNimbusFeatures()
+        }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        PreferenceManager.getDefaultSharedPreferences(this)
+            .unregisterOnSharedPreferenceChangeListener(SettingsListener)
+    }
+}
+
+fun Context.showCustomDialog(message: String): AlertDialog = AlertDialog.Builder(this)
+    .setCancelable(false)
+    .create().apply {
+        setView(CustomDialogBinding.inflate(LayoutInflater.from(this@showCustomDialog)).apply {
+            description.text = getString(R.string.custom_dialog_message, message)
+            button.setOnClickListener { dismiss() }
+        }.root)
+    }
 
 class NavigationFragment : Fragment() {
